+++ src/pages/use.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Head from 'next/head';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { ILocale } from '@/interfaces/locale';
-import Header from '@/components/common/header_1';
+import Header from '@/components/common/header';
 import UsePageBody from '@/components/use_page/use_page_bode';
 import Footer from '@/components/common/footer';
